refactor and chore: adding functions from darias file
