+++ lexer.h
@@ -10,8 +10,8 @@ enum token_type
     RED_TO,     // 3
     RED_FROM,   // 4
     APPEND,     // 5
-    HERE_DOC,    // 6
-    ASSIGNMENT_WORD
+    HERE_DOC,   // 6
+    ASSIGNMENT_WORD, // to delete when possible
 };
 
 typedef struct s_token
@@ -39,9 +39,16 @@ typedef struct s_parse_tree
 #define SUBTREE_OK 0
 #define MEMORY_ERROR 1
 
-// Parsing
+// Allocs and frees
 t_parse_tree* alloc_parse_tree();
 void free_parse_tree(t_parse_tree *tree);
+void free_token(t_token* tok);
+void free_token_list(t_token_list* list);
+
+// Errors
+void handle_memory_error(t_token **token_list, int num_tokens);
+
+// Parsing
 void link_node(t_parse_tree **current, t_parse_tree *newNode);
 int is_cmd_word(t_token_list **tok, t_parse_tree **cmd_word_node);
 int is_simple_command(t_token_list **tok, t_parse_tree **new);
@@ -56,18 +63,20 @@ int is_pipe_sequence(t_token_list **tok, t_parse_tree **new);
 void print_parse_tree(t_parse_tree* tree, int depth);
 
 // Testing
-void test_parser();
+void test_parser(); // to delete later
 
 // Tokens
-void free_token(t_token* tok);
-t_token_list* create_token(enum token_type type, char* lexeme); //see if it is necessary to include
+t_token_list* create_token(enum token_type type, char* lexeme); //probably to delete later
 void add_token(t_token_list** list, enum token_type type, char* lexeme); // same as above
-void free_token_list(t_token_list* list);
 
 // Lexer helper
 int count_additional_chars(char *input, const char *delim);
 char *preprocess_input(char *str, const char *delim);
-
+size_t handle_quotes(char *str, int *error);
+char	*ft_strtok(char *str, const char *delim, int *error); // Maybe add to Libft
+int	ft_strlen(const char *s); // same as above
+char *update_pointer(char *str); // same?
 
 // Lexer
-void lexer(char *input, t_token_list **tokenList, int *numTokens);
+void lexer(char *input, t_token_list **tokenList, int *numTokens, int *error);
+enum token_type determine_token_type(char *token_value);
